Display data user
from database to the UI
